+++ app/controllers/TipoModulo.php
@@ -19,7 +19,8 @@ class TipoModulo extends Controller{
     }
 
     public function create(){
-        if ($_SERVER['REQUEST_METHOD'] == 'POST'){
+        if (($_SERVER['REQUEST_METHOD'] == 'POST')&&$this->sessionActiva())
+        {
             $datos = [
                 'nombre' => trim($_POST['nombreTipoModulo']),
                 'estado' => trim($_POST['estado'])
@@ -30,16 +31,18 @@ class TipoModulo extends Controller{
                 die("Error al insertar los datos");
             }
         }else{
-            $datos = [
-                'nombre' => '',
-                'estado' => ''
-            ];
-            $this->view('pages/tipoModulo', $datos);
+            // $datos = [
+            //     'nombre' => '',
+            //     'estado' => ''
+            // ];
+            // $this->view('pages/tipoModulo', $datos);
+            redireccionar('tipoModulo');
         }
     }
 
     public function update(){
-        if ($_SERVER['REQUEST_METHOD'] == 'POST'){
+        if (($_SERVER['REQUEST_METHOD'] == 'POST')&&$this->sessionActiva())
+        {
             $datos = [
                 'id_tipo_modulo' => $_POST['id_tipo_modulo'],
                 'nombre' => trim($_POST['nombreTipoModulo']),
@@ -56,22 +59,26 @@ class TipoModulo extends Controller{
         }
     }
 
-    public function delete($id){
-        if (isset($id)){
+    public function delete($id=null){
+        if(isset($id)&&$this->sessionActiva())
+            {
             if($this->tipoModuloModel->delete($id)){
                 // redireccionar('tipoModulo');
             }else{
+                redireccionar('tipoModulo');
                 die("Error al eliminar los datos");
             }
         }else{
-            // $this->index();
+             $this->index();
         }
     }
 
-    public function  findById($id){
-        if(isset($id)){
+    public function  findById($id = null){
+        if(isset($id)&&$this->sessionActiva())
+            {
             return $this->tipoModuloModel->findById($id);
         }else{
+            redireccionar('tipoModulo');
             die("Error al buscar el dato");
         }
     }
